[Security Solution][Alerts] saved query UX changes follow-up (#141747)
## Summary
addresses feedback left during code review for https://github.com/elastic/kibana/pull/140064

- Saved query checkbox label shows saved query name
- `createSavedQuery` moved to `x-pack/plugins/security_solution/cypress/tasks/api_calls/saved_queries.ts`
- `useGetSavedQuery` moved to `x-pack/plugins/security_solution/public/detections/pages/detection_engine/rules/details`
- additional comments and minor code changes

### Before
<img width="762" alt="Screenshot 2022-09-26 at 11 38 05" src="https://user-images.githubusercontent.com/92328789/192256482-2bfc29c9-305a-4ff6-b9cd-b0505e887bd1.png">

### After
<img width="989" alt="Screenshot 2022-09-26 at 11 37 23" src="https://user-images.githubusercontent.com/92328789/192256378-290183b1-44e2-47bb-9d64-c46d0819cff4.png">

(cherry picked from commit f585dd6e5a3d0336d2a97fd0c5ee3ff88ccf0ab0)
