+++ x-pack/plugins/security_solution/cypress/tasks/api_calls/saved_queries.ts
@@ -33,3 +33,20 @@ export const createSavedQuery = (
     },
     headers: { 'kbn-xsrf': 'cypress-creds' },
   });
+
+export const deleteSavedQueries = () => {
+  const kibanaIndexUrl = `${Cypress.env('ELASTICSEARCH_URL')}/.kibana_\*`;
+  cy.request('POST', `${kibanaIndexUrl}/_delete_by_query?conflicts=proceed`, {
+    query: {
+      bool: {
+        filter: [
+          {
+            match: {
+              type: 'query',
+            },
+          },
+        ],
+      },
+    },
+  });
+};
