+++ x-pack/plugins/security_solution/cypress/tasks/common.ts
@@ -179,23 +179,6 @@ export const deleteCases = () => {
   });
 };
 
-export const deleteSavedQueries = () => {
-  const kibanaIndexUrl = `${Cypress.env('ELASTICSEARCH_URL')}/.kibana_\*`;
-  cy.request('POST', `${kibanaIndexUrl}/_delete_by_query?conflicts=proceed`, {
-    query: {
-      bool: {
-        filter: [
-          {
-            match: {
-              type: 'query',
-            },
-          },
-        ],
-      },
-    },
-  });
-};
-
 export const postDataView = (dataSource: string) => {
   cy.request({
     method: 'POST',
