+++ x-pack/plugins/security_solution/public/detections/components/rules/step_define_rule/translations.tsx
@@ -78,12 +78,21 @@ export const EQL_QUERY_BAR_LABEL = i18n.translate(
   }
 );
 
-export const SAVED_QUERY_CHECKBOX_LABEL = i18n.translate(
-  'xpack.securitySolution.detectionEngine.createRule.stepDefineRule.SavedQueryCheckboxLabel',
+export const SAVED_QUERY_FORM_ROW_LABEL = i18n.translate(
+  'xpack.securitySolution.detectionEngine.createRule.stepDefineRule.SavedQueryFormRowLabel',
   {
     defaultMessage: 'Saved query',
   }
 );
+
+export const getSavedQueryCheckboxLabel = (savedQueryName: string) =>
+  i18n.translate(
+    'xpack.securitySolution.detectionEngine.createRule.stepDefineRule.fieldShouldLoadQueryDynamicallyLabel',
+    {
+      defaultMessage: 'Load saved query "{savedQueryName}" dynamically on each rule execution',
+      values: { savedQueryName },
+    }
+  );
 
 export const THREAT_MATCH_INDEX_HELPER_TEXT = i18n.translate(
   'xpack.securitySolution.detectionEngine.createRule.stepDefineRule.threatMatchingIcesHelperDescription',
